initial commit of mwm_legacy_ir2opt.py
